* sanity.sh: added "cat > /dev/null" to loginfo entry to avoid the 	  SIGPIPE signal
	* patch.c: added new variable diff_opt
	(patch): allow optional parameter to -c and -u option, send it to
	server
	(patch_fileproc): cleaned up the code which prints the current
	filename.  For "-s" option, print the pathname relative to CVSROOT
	instead of just the filename.

	* filesubr.c (xchmod): added cast to shut up gcc

	* cvs.h: added prototype for gethostname

	* lock.c (write_lock), (Reader_Lock), import.c (update_rcs_file),
	  client.c (update_entries), (send_modified), server.c (server),
	  (receive_file), (server_updated): use %ld for printing pid_t
	  variables

CVS-ID: src/ChangeLog 1.493
CVS-ID: src/client.c 1.101
CVS-ID: src/cvs.h 1.74
CVS-ID: src/filesubr.c 1.14
CVS-ID: src/import.c 1.47
CVS-ID: src/lock.c 1.23
CVS-ID: src/logmsg.c 1.19
CVS-ID: src/patch.c 1.26
CVS-ID: src/sanity.sh 1.61
CVS-ID: src/server.c 1.67

+++ src/filesubr.c
@@ -325,9 +325,11 @@ xchmod (fname, writable)
     if (trace)
 #ifdef SERVER_SUPPORT
 	(void) fprintf (stderr, "%c-> chmod(%s,%o)\n",
-			(server_active) ? 'S' : ' ', fname, mode);
+			(server_active) ? 'S' : ' ', fname,
+			(unsigned int) mode);
 #else
-	(void) fprintf (stderr, "-> chmod(%s,%o)\n", fname, mode);
+	(void) fprintf (stderr, "-> chmod(%s,%o)\n", fname,
+			(unsigned int) mode);
 #endif
     if (noexec)
 	return;
